[+] 增加 后台管理密码校验接口 [*] 优化 `BDWPApiController` `isdir` 字段为 `is_dir`

+++ config/mail.php
@@ -45,7 +45,7 @@ return [
         'smtp' => [
             'transport' => 'smtp',
             'host' => env('MAIL_HOST', '127.0.0.1'),
-            'port' => env('MAIL_PORT', 2525),
+            'port' => (int)env('MAIL_PORT', 2525),
             'encryption' => env('MAIL_ENCRYPTION', 'tls'),
             'username' => env('MAIL_USERNAME'),
             'password' => env('MAIL_PASSWORD')
@@ -70,6 +70,6 @@ return [
 
     'to' => [
         'address' => env('MAIL_TO_ADDRESS', 'hello@example.com'),
-        'name'    => env('MAIL_TO_NAME', 'Example'),
+        'name' => env('MAIL_TO_NAME', 'Example'),
     ]
 ];
